Generate and block listing of photos until thumbnails are available
The generation job times out after 30 seconds in queue. It is assumed
that it should complete in well under that time and checks for
existence of the job or the resulting file are in place so it should
not DoS worker or slow down navigation.

This should be ultimately moved to a separate WSGI application like
iw.thumbs, repoze.bitblt or wsgithumbs but current state of these
application is not satisfactory.